pyproject.toml
[project]
name = "GDG-Hackathon-ML6-Agent"
version = "0.1.0"
description = "This repository holds the scaffolding and dummy agent for the GDGxML6 Hackathon."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "pandas",
    "matplotlib",
    "seaborn",
    "colorama>=0.4.6",
    "pyfiglet>=1.0.4",
    "PyMuPDF",
    "beautifulsoup4>=4.14.3",
    "ddgs",
    "python-chess>=1.999",
    "stockfish>=4.0.8",
]
